src/registry_rest_api/func/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi==0.68.1
fastapi-pagination==0.8.3
fastapi-utils==0.2.1
requests==2.26.0
